refactor: kebab-case naming convention
