+++ src/components/app-header/app-header.jsx
@@ -1,5 +1,5 @@
 import {Logo, BurgerIcon, ListIcon, ProfileIcon} from "@ya.praktikum/react-developer-burger-ui-components";
-import headerStyles from './appHeader.module.css';
+import headerStyles from './app-header.module.css';
 
 
 function AppHeader() {
