+++ src/components/app/app.jsx
@@ -1,6 +1,6 @@
 import styles from "./app.module.css";
 import { data } from "../../utils/data";
-import AppHeader from "../appHeader/appHeader";
+import AppHeader from "../app-header/app-header";
 
 function App() {
   return (
